+++ RadCars.Web.ViewModels/CarMake/CarMakeViewModel.cs
@@ -2,7 +2,7 @@
 
 public class CarMakeViewModel
 {
-    public ushort Id { get; set; }
+    public int Id { get; set; }
 
     public string Name { get; set; } = null!;
 }
